refactor(LottoTicket): 등수 체크 책임을 LottoTicket으로 이동
- LottoTicketChecker에서 수행하던 로또 등수 체크 로직을 LottoTicket로 이전
- updateRank 메서드 사용하지 않고, LottoTicket 내부에서 스스로 등수 계산
- LottoTicketChecker 클래스 삭제

외부에서 LottoTicket의 내부 상태를 직접 수정(updateRank)하는 구조는
캡슐화가 깨져 개선이 필요하다고 생각했다.

LottoTicket이 스스로 자신의 등수를 계산하도록 하여
객체지향 원칙에 맞게 책임을 이동시켰다.
