+++ src/main/java/lotto/model/LottoTicketGenerator.java
@@ -1,7 +1,5 @@
 package lotto.model;
 
-import lotto.model.lottoTicket.LottoTicket;
-
 public class LottoTicketGenerator {
     public static LottoTicket generateLottoTicket(Lotto lotto) {
         return new LottoTicket(lotto);
